+++ cbits/proAudioSdl.h
@@ -84,6 +84,8 @@ public:
     virtual void soundStop();
     /// returns number of currently active sounds
     virtual unsigned int soundActive() const;
+    /// check if a specified sound is active
+    virtual bool soundActive(uint64_t sound);
 protected:
     /// constructor. Use the create() method instead
     DeviceAudioSdl(unsigned int nTracks, unsigned int frequency, unsigned int chunkSize);