+++ cbits/proteaaudio_binding.h
@@ -16,7 +16,7 @@ sample_t _sampleFromMemoryWav(char *data, int size, float volume);
 sample_t _sampleFromMemoryOgg(char *data, int size, float volume);
 sample_t sampleFromFile(char* filename, float volume);
 int sampleDestroy(uint64_t sample);
-int soundActive();
+int soundActiveAll();
 void soundStopAll();
 sound_t soundLoop(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
 sound_t soundPlay(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
@@ -24,6 +24,7 @@ sound_t soundPlayOn(unsigned int track, sample_t sample, float volumeL, float vo
 sound_t soundLoopOn(unsigned int track, sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
 int soundUpdate(sound_t sound, float volumeL, float volumeR, float disparity, float pitch);
 int soundStop(sound_t sound);
+int soundActive(sound_t sound);
 #ifdef __cplusplus
 }
 #endif